Create Round Robin Controller.vhd
There are three microprocessors A, B and C and are required to share the access to the same synchronous. Each microprocessor will initiate a RAM request signal when it wants access to the RAM and will deactivate it when finished. If more than of the microprocessors request access at the same time, the access is going to be giving by “Round-Robin” basis.