Create interrupts.c
For this lab we learned how to create a new Zynq hardware which included interrupt components and be able to mix our knowledge that we gained from lab two into this new lab. One of the GPIO uses push buttons to create the interrupts and the other GPIO will be the eight LEDs. The two GPIO are going to be connected to the process system through AXI interfaces.